+++ example/lib/screens/home.dart
@@ -81,7 +81,7 @@ class _ExampleAlarmHomeScreenState extends State<ExampleAlarmHomeScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: const Text('alarm 4.0.0-dev.3')),
+      appBar: AppBar(title: const Text('alarm 4.0.0-dev.4')),
       body: SafeArea(
         child: alarms.isNotEmpty
             ? ListView.separated(